\name{geneplast-package}
\alias{geneplast-package}
\alias{geneplast}
\docType{package}
\title{
Geneplast: an R package for evolutionary rooting and plasticity inference based on orthologous groups distribution.
}

\description{
Geneplast is designed for evolutionary and plasticity analysis based on orthologous groups distribution in a given species tree. It uses Shannon information theory and orthologs abundance to estimate the Evolutionary Plasticity Index. Additionally, it implements the Bridge algorithm to determine the evolutionary root of a given gene based on its orthologs distribution.
}

\details{
\tabular{ll}{
Package: \tab geneplast\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-06-10\cr
License: \tab GPL (>= 2)\cr
}
R package for gene plasticity inference based on orthologous groups distribution.
}

\author{
Rodrigo JS Dalmolin, Mauro AA Castro.
}

\references{
Dalmolin RJS et al. Geneplast: an R package for gene plasticity inference based on orthologous groups distribution. Journal paper (in preparation).
}

\keyword{ package }
